\name{RemoteREngine.test-package}
\alias{RemoteREngine.test-package}
\alias{RemoteREngine.test}
\docType{package}
\title{
Test suite for RemoteREngine
}
\description{
This package contains unit test suite for the RemoteREngine package
}
\details{
\tabular{ll}{
Package: \tab RemoteREngine.test\cr
Type: \tab Package\cr
Version: \tab 0.0-0\cr
Date: \tab 2009-09-16\cr
License: \tab GPL (>= 3.0)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Romain Francois <francoisromain@free.fr>

Maintainer: Romain Francois <francoisromain@free.fr>
}
\keyword{ package }
